json-api/requirements.txt
kafka-python
uvicorn